Factor out logic to create all base ballot props
